# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: North API
  version: 1.0.0
  description: |
    JSON API for North's native clients. The iOS app generates its client from
    this file (swift-openapi-generator), so this file is the contract: a route
    under /api/v1 that is not described here fails cmd/web's contract test.

    Authentication, by route group:
    - `/auth/*` sign-in routes take no credentials and return a session token.
    - Everything else takes that session token as `Authorization: Bearer`.
    - `/capture/*` takes an `nk_` connection token instead, issued from
      Settings → Connections for agents and shortcuts.

    A browser session cookie authenticates none of these.
servers:
  - url: https://kheprios.com/api/v1
    description: Production
  - url: http://localhost:8090/api/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /day:
    get:
      operationId: getDay
      summary: My Day — one local date across every card
      parameters:
        - name: date
          in: query
          required: false
          description: YYYY-MM-DD in the person's zone. Defaults to today; a malformed value is today.
          schema:
            type: string
            format: date
      responses:
        '200':
          description: The day
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /day/rules:
    get:
      operationId: listDayRules
      summary: The standing rules a day is read against (kitchen closes, screens off…)
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayRulesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /day/rules/{kind}:
    parameters:
      - name: kind
        in: path
        required: true
        description: caffeine_cutoff, kitchen_closes, last_drink or screens_off.
        schema:
          type: string
    put:
      operationId: setDayRule
      summary: Create or replace one rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DayRuleInput'
      responses:
        '200':
          description: Every rule, after the change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayRulesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteDayRule
      summary: Remove one rule
      responses:
        '200':
          description: Every rule, after the change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayRulesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token returned by the /auth routes.
    connectionToken:
      type: http
      scheme: bearer
      description: An `nk_` connection token from Settings → Connections.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    DayCaffeine:
      type: object
      required: [totalMg, activeMg, limitMg, afterCutoff]
      properties:
        totalMg:
          type: integer
        activeMg:
          type: integer
        limitMg:
          type: integer
        afterCutoff:
          type: boolean
    DayFast:
      type: object
      required: [startedAt, targetHours, elapsedMinutes, phase, fraction]
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        targetHours:
          type: integer
        elapsedMinutes:
          type: integer
        phase:
          type: string
          enum: [fed, fasting, fat_burning, ketosis]
        fraction:
          type: number
    DayNutrients:
      type: object
      required: [covered, missing, total]
      properties:
        covered:
          type: array
          items:
            type: string
        missing:
          type: array
          items:
            type: string
        total:
          type: integer
    DayMilestone:
      type: object
      required: [id, name, monthsSince, fraction, due]
      properties:
        id:
          type: string
        name:
          type: string
        monthsSince:
          type: integer
        fraction:
          type: number
        due:
          type: boolean
    DayBloodPressure:
      type: object
      required: [systolic, diastolic, at]
      properties:
        systolic:
          type: integer
        diastolic:
          type: integer
        at:
          type: string
          format: date-time
    DaySoreness:
      type: object
      required: [region, severity]
      properties:
        region:
          type: string
        severity:
          type: integer
    DayResponse:
      type: object
      required: [date, isToday, now, vitals, food, water, activity, workouts, body, streak, level, caffeine, nutrients, milestones, timeline, markers]
      properties:
        date:
          type: string
          format: date
        isToday:
          type: boolean
        now:
          type: string
          format: date-time
        vitals:
          $ref: '#/components/schemas/DayVitals'
        food:
          $ref: '#/components/schemas/DayFood'
        water:
          $ref: '#/components/schemas/DayWater'
        activity:
          $ref: '#/components/schemas/DayActivity'
        sleep:
          $ref: '#/components/schemas/DaySleep'
        workouts:
          $ref: '#/components/schemas/DayWorkouts'
        body:
          $ref: '#/components/schemas/DayBody'
        streak:
          type: integer
        level:
          type: integer
        caffeine:
          $ref: '#/components/schemas/DayCaffeine'
        fast:
          $ref: '#/components/schemas/DayFast'
        nutrients:
          $ref: '#/components/schemas/DayNutrients'
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/DayMilestone'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/DayTimelineEntry'
        markers:
          type: array
          items:
            $ref: '#/components/schemas/DayMarker'
    DayVitals:
      type: object
      description: Absent fields were not measured.
      properties:
        energyPercent:
          type: integer
        daylightMinutes:
          type: integer
        screenMinutes:
          type: integer
    DayFood:
      type: object
      required: [calories, proteinG, carbG, fatG]
      properties:
        calories:
          type: number
        proteinG:
          type: number
        carbG:
          type: number
        fatG:
          type: number
        goal:
          $ref: '#/components/schemas/DayMacroGoal'
    DayMacroGoal:
      type: object
      required: [calories, proteinG, carbG, fatG]
      properties:
        calories:
          type: number
        proteinG:
          type: number
        carbG:
          type: number
        fatG:
          type: number
    DayWater:
      type: object
      required: [totalMl, targetMl]
      properties:
        totalMl:
          type: integer
        targetMl:
          type: integer
    DayRing:
      type: object
      required: [value, goal, percent]
      properties:
        value:
          type: number
        goal:
          type: number
        percent:
          type: integer
    DayActivity:
      type: object
      required: [move, exercise, stand]
      properties:
        move:
          $ref: '#/components/schemas/DayRing'
        exercise:
          $ref: '#/components/schemas/DayRing'
        stand:
          $ref: '#/components/schemas/DayRing'
    DaySleepBlock:
      type: object
      required: [stage, start, end]
      properties:
        stage:
          type: string
          enum: [deep, rem, core, awake]
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    DaySleep:
      type: object
      required: [totalMinutes, stages, blocks, source]
      properties:
        totalMinutes:
          type: integer
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        stages:
          type: object
          description: Minutes per stage (deep, rem, core, awake). Empty for a manual log.
          additionalProperties:
            type: integer
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/DaySleepBlock'
        quality:
          type: integer
        source:
          type: string
    DayWorkouts:
      type: object
      required: [count, minutes, calories, labels]
      properties:
        count:
          type: integer
        minutes:
          type: integer
        calories:
          type: number
        labels:
          type: array
          items:
            type: string
    DayBody:
      type: object
      required: [soreness]
      properties:
        weightKg:
          type: number
        heightCm:
          type: number
        bmi:
          type: number
        bmiCategory:
          type: string
          enum: [underweight, healthy, overweight, obese]
        targetWeightKg:
          type: number
        toGoalKg:
          type: number
        bloodPressure:
          $ref: '#/components/schemas/DayBloodPressure'
        soreness:
          type: array
          items:
            $ref: '#/components/schemas/DaySoreness'
    DayTimelineEntry:
      type: object
      required: [kind, at, title, icon]
      properties:
        kind:
          type: string
        at:
          type: string
          format: date-time
        title:
          type: string
        detail:
          type: string
        href:
          type: string
        icon:
          type: string
    DayMarker:
      type: object
      required: [kind, label, at, passed]
      properties:
        kind:
          type: string
        label:
          type: string
        at:
          type: string
          format: date-time
        passed:
          type: boolean
    DayRule:
      type: object
      required: [kind, label, at, enabled]
      properties:
        kind:
          type: string
        label:
          type: string
        at:
          type: string
          description: HH:MM
        enabled:
          type: boolean
    DayRulesResponse:
      type: object
      required: [rules, kinds]
      properties:
        rules:
          type: array
          items:
            $ref: '#/components/schemas/DayRule'
        kinds:
          type: array
          items:
            type: string
    DayRuleInput:
      type: object
      required: [at, enabled]
      additionalProperties: false
      properties:
        at:
          type: string
          description: HH:MM
        enabled:
          type: boolean
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [message]
      properties:
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...